Model/requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
trl>=0.7.4
peft>=0.7.0
bitsandbytes>=0.41.0
huggingface_hub>=0.19.0
hf_transfer>=0.1.0
accelerate>=0.24.0
tensorboard>=2.14.0
pillow>=10.0.0
pandas>=1.5.0
